spec/13-implementation/tools/requirements.local.txt
pyyaml>=6.0
jsonschema>=4.22.0
psycopg2-binary>=2.9.9

